requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.27
aiosqlite==0.19.0 